Add constraint on Chef to not support version >= 17
With Chef 17, the knife code has been extracted to a dedicated gem.
=> This means that to be compatible with Chef 17 we should leverage the
knife gem which only supports Chef 17.

As for now we want to be compatible with older Chef version we can't do
that so easily.

The idea is to create a 1.6 branch which support Chef until version 17,
and once stable [1], release a 2.0 version of cookbook release with only
Chef 17 support.

[1] otherwise we'll have to backport fixes on 1.6 branch