Fix git root directory discovery
Initial implementation could produce invalid result when the code is
evaluated in a git directory which is not related with the cookbook.

For instance when berkshelf fetches cookbooks from git source it will
checkout the code in a configured berkshelf_path directory and evaluate
the metadata.rb of the cookbook.
If the configured berkshelf_path is inside a git directory, this will
be considered as the "git root dir" of the cookbook and will mess up
with the version computation.

This commit fixes the GitUtilities.find_root logic by quickly checking
whether the given directory is tracked by git.
If the dir is not tracked or the command used to determine the tracking
status just fail, it means it is not part of the parent git repository.